[ch-fix-styling-issues-162417481] refactor code to reduce complexity
